Fix flaky test (#105)
@istzen fixing this test because it is pretty annoying in the pipeline
when it intermittently fails :D

The main problem was the missing `setUpAll`. That meant that the code
was just placed inside the group and influenced the other tests as far
as I could figure out (not 100% sure what exactly was going on for the
failed tests, but it's definitely wrong). For example, in line 53 there
should not have been any ride present because it was never initialized??
It's just lucky that the tests worked so far.