Better PasswordField (User study feedback) (#100)
- Main thing: Add a toggle button for obscuring the password. This also
required the widget to become stateful, of course.
- Refactor the PasswordField because the validator logic was always one
of three things. Now the pages just need to say whether they want to
validate strictly (Reset + Register vs. Login) and if the password is a
confirmation field (by providing the controller of the original password
field)
- Adds a HelperText for the strictly validating fields to make the
requirements more clear. It actually updates on change and always tells
you the next requirement to fulfill.
- Refactor the tests into their own file because there was a lot of
duplication (only possible now that the logic actually resides inside
the widget but I remember @istzen asking for it on the original PR)

I also saw that it is possible to do form validation on user input, is
that something we should maybe do?