Settings test (#111)
tests for the accountPage (SettingsPage in the Future). 
I did not test the launchUrlString method in the HelpPage/AboutPage we
can do that later. And since there are no other relevant things to test
in these Pages i just left them out.
Also i moved the methods of the AccountPage below the build method as in
the other pages.

---------

Co-authored-by: Philipp <philipp.lerche@sicher3.de>