Enable very strict type checks (#104)
This enables the strictest type checks that dart has to offer. It
concerns mainly the `fromJson` methods where we need to cast the values
before applying them.

Also, the `select` method needs to be cast to the right return type
(depending on whether there is a `.single()`: List<Map<String,dynamic>>
or just Map<String, dynamic>).