Tappable editable row (#123)
As suggested in the user study, where almost everyone expected the
respective category titles to be tappable. I'm not quite sure yet if
really the whole row should be tappable or if we should keep the icon
buttons and just make the text clickable as well (like for the
username). Feedback appreciated!
@Showcas This could have a slight impact on your tests.

---------

Co-authored-by: Fynn Starke <f.starke@iot-venture.com>