Recurring drives (#128)
This is basically just the backend for later (SQL and Flutter models).

Please review especially the ideas in here. Especially the sql functions
are pretty important.

One thing that you cannot see in the dump: We added a cronjob that runs
every night at 3 am and starts the create_drives_from_recurring
function. That one creates a new drive in 30 days, if required by the
recurrence rule.