Test search ride (#114)
Search Ride and Search Ride Filter are 100% tested*, and I kept the test
file just under 1000 lines!

* I found one (VERY minor) with the spacing in the indicator row.
* I refactored the swap start and destination button into the
StartDestinationTimeline.
* The price is now actually dependent on the distance. I needed that in
order to test the ordering by price. Because we want to test every
function we add from now on, I also tested the Position model.
* I added a goodie: If you look for a ride and there is one present but
not at the time you selected, it will tell you to try another time, now
(instead of the "Relax your filters" text, which I noticed while testing
wasn't really fitting). If you tap on the text, it will automatically
take you to the next-best time.

\* If you test this between 00:00 and 00:10, there is one line that
isn't tested. The line in question is the validator telling you to
choose a valid time, as you chose a time more than 10 minutes ago.
Because we use DateTime.now() in that test and I didn't find a good way
to mock that online, I thought I'd leave it as is for now. The internet
recommends using dart.clock if you want to mock the time afaik.

---------

Co-authored-by: Fynn Starke <f.starke@iot-venture.com>