Fix username hint (#145)
Just a very quick PR based on Felix' feedback that he was unable to see
the whole hint text.

![Screenshot_20230302-163455](https://user-images.githubusercontent.com/63241108/222949382-adb9f9d2-9332-401e-9771-95f9878f351a.png)

The hint was very inconsistent anyways. So I moved that to the helper
text below (does not take more space because there is already the 0/15
display because of the maximal length) and wrote a hint consistent to
what we normally write. With that one, it does not matter if it is cut
off because it is not vital information.