Redesign home page (#142)
As promised, I redesigned the home page!
* I divided the "notifications" area into 3 parts. To do that, I split
the all-encompassing `_items` list into what I consider its natural 3
parts (upcoming trips, ride events, and messages). I also reordered them
in the code to reflect what will be shown on screen, so that's probably
the majority of what you see in the diff :D
* I added a new notification: Complete profile! It tells the user that
they're "very mysterious" and there is a button to the profile page. It
appears if no other notifications are present (If the user has "nothing
else to do"), and the user hasn't configured any personal info except
features (I didn't consider features personal information, but we can
debate that). I am aware that the page doesn't update if you change your
profile via the normal way. I think if we want to implement that, we'll
have to add a listener and notify them every time currentProfile
changes. I didn't consider that vital, but if you want me to do that,
let me know :)