Redesign trip overview (#143)
You can consider this Part I of the Detail Pages redesign. It's kind of
more of a personal project because I always considered this the very
ugliest part of the app (I particularly disliked how small and
disconnected start and destination were: You'd probably never notice it
at all since there are bigger, more colorful elements underneath - and
if you did, you'd not get the connection right away).

I tried to fix that for the Drive Overview by making start and end
bigger and bolder while making the times less bold and redesigning the
rest of the info to be around, not strictly underneath that.

And yes, you heard correctly: There is now a Drive Overview, which means
there is also a Ride Overview. The reason for the start-destination part
of the overview not being a timeline is that we felt it looked weird on
the Drive Detail Page (where there is a basically identical timeline
underneath it, containing the stops and where to pick people up and drop
them off). However, I split the Overviews because I still think it makes
more sense on the Ride Detail Page to include a proper timeline, like
you would see on the cards. If you don't like it, I can revert that.
Also, I included the price in the Ride Overview.

Speaking of cards, I redesigned them just a smidge to make them
consistent with the boldness and sizes of the start and destination
texts of the Drive Overview. A side effect is that they now contain just
a little less dead space, which was intentional.

---------

Co-authored-by: Fynn Starke <f.starke@iot-venture.com>