Redesign drive detail page (#144)
Finally, a redesign of the drives detail page! I've never been a fan of
the containers with a border, and thought it much more practical to use
what I think Flutter intends for this use case: A card. I think this
worked out quite well. Otherwise, I made a few changes to font sizes,
spacing etc here and there.

It is worth mentioning that I included a few headers like "Driver" on
the Ride Detail Page or "Route" on the Drive Detail Page. I think this
looks better than Dividers. It is also a response to the user studies,
where some users didn't get that the features belonged to the drivers,
or why the start and destination information is displayed twice on the
Drive Detail Page.