Recurring drives frontend (#146)
Redoing #135 which I accidentally merged.

We're truly sorry, this is the biggest feature by far. This is the front
end part of the recurring drives!
Instead of all the new changes, I'm going to describe a flow for
testing:

You can create recurring drives by clicking the checkbox on the regular
create drive page. Please check if everything is understandable for
usability.
You are taken to the new recurring drive detail page, where you can see
your schedule, information and all generated drives. In addition to
that, you can edit the schedule by tapping the icon on the top right.
You can see your recurring drives on the new "Recurring" tab under the
Drives navigator bar entry.
The rest of the time was essentially spent on figuring out edge cases
and weirdness that has to do with the complexity of situations that can
arise by cancelling, changing or picking weird recurring drive
schedules.

---------

Co-authored-by: Simon Hock <simon.hock@stud.tu-darmstadt.de>