Test accessibility (#150)
These are automated tests for accessibility. It checks:
* Whether every tappable object is big enough: That is, at least 48*48
pixels.
* Whether every tappable object has a label or tooltip for Screen
Readers.
* Whether the WCAG recommended contrast ratio of at least 4.5/1 is met.
For more information, look
[here](https://www.w3.org/TR/UNDERSTANDING-WCAG20/visual-audio-contrast-contrast.html).

I had to change a few things in order to get this to work: 
* First of all, the Write Review Page looks a lot different now. Love it
or hate it, but I needed to change it so the tap target guideline was
met and at least, now the category texts don't wrap nastily anymore.
* I had to remove the InkWell from the values on the ProfilePage, as
they weren't big enough either. I tried making them 48 pixels high, but
that looked very bad. Also, I made the tap targets of the change avatar
button and the username bigger (the latter results in the edit icon
always being on the far right).
* I labeled everything tappable.
* Sometimes, I replaced `...onSurface.withOpacity(0.5)` with
`...onSurface.withOpacity(0.6)` in order to meet the contrast guideline.

---------

Co-authored-by: Fynn Starke <f.starke@iot-venture.com>