Last bug fixes (#158)
This fixes a few very minor bugs we discovered during and after the
presentation:

* Reviews were ordered incorrectly. For the order it was shown in, we
only took into account the creation date of the review, not when it was
last updated.
* When looking for a start and destination that there are rides for, but
not at the time you were looking for, you can tap the text in order to
automatically go to the next best possible time. In order to do this, we
take the time of the first ride in your selected order. Because it's
usually not possible to order your search results by time proximity when
searching on a whole day, this would lead to the time proximity not
being taken into account if the default order "relevance" was selected.
This lead to you probably not being taken to the nearest possible time.
* We tweaked the search system slightly: Previously, we searched the
database for matching drives, then created a preview of a ride per
drive. Now, we take into account whether you already have rides with the
drives: If you have a pending or accepted ride, this ride will be shown
directly instead of a preview, and if your ride request was rejected, a
ride won't be shown at all.
* We removed the CODEOWNERS file.